# frontend/src/api/openapi.json
{
  "components": {
    "schemas": {
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "LoginRequest": {
        "additionalProperties": false,
        "properties": {
          "email": {
            "maxLength": 254,
            "minLength": 1,
            "title": "Email",
            "type": "string"
          },
          "password": {
            "maxLength": 128,
            "minLength": 1,
            "title": "Password",
            "type": "string"
          }
        },
        "required": [
          "email",
          "password"
        ],
        "title": "LoginRequest",
        "type": "object"
      },
...
      "SessionResponse": {
        "additionalProperties": false,
        "properties": {
          "absoluteExpiresAt": {
            "title": "Absoluteexpiresat",
            "type": "string"
          },
          "displayName": {
            "title": "Displayname",
            "type": "string"
          },
          "idleExpiresAt": {
            "title": "Idleexpiresat",
            "type": "string"
          },
          "role": {
            "enum": [
              "admin",
              "agent"
            ],
            "title": "Role",
            "type": "string"
          },
          "userId": {
            "title": "Userid",
            "type": "string"
          }
        },
        "required": [
          "userId",
          "displayName",
          "role",
          "idleExpiresAt",
          "absoluteExpiresAt"
        ],
        "title": "SessionResponse",
        "type": "object"
      },
...
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      }
    }
  },
  "info": {
    "title": "Real-Estate Lead Capture CRM",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
    "/api/v1/auth/login": {
      "post": {
        "operationId": "login",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/LoginRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SessionResponse"
                }
              }
            },
            "description": "Successful Response",
            "headers": {
              "Cache-Control": {
                "description": "Sensitive API responses are private and not stored",
                "schema": {
                  "type": "string"
                }
              },
              "X-Request-ID": {
                "description": "Opaque request correlation identifier",
                "schema": {
                  "format": "uuid",
                  "type": "string"
                }
              }
            }
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error",
            "headers": {
              "Cache-Control": {
                "description": "Sensitive API responses are private and not stored",
                "schema": {
                  "type": "string"
                }
              },
              "X-Request-ID": {
                "description": "Opaque request correlation identifier",
                "schema": {
                  "format": "uuid",
                  "type": "string"
                }
              }
            }
          }
        },
        "summary": "Login",
        "tags": [
          "authentication"
        ]
      }
    },
    "/api/v1/auth/logout": {
      "post": {
        "operationId": "logout",
        "parameters": [
          {
            "in": "header",
            "name": "X-CSRF-Token",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Csrf-Token"
            }
          },
          {
            "in": "cookie",
            "name": "session_id",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Session Id"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response",
            "headers": {
              "Cache-Control": {
                "description": "Sensitive API responses are private and not stored",
                "schema": {
                  "type": "string"
                }
              },
              "X-Request-ID": {
                "description": "Opaque request correlation identifier",
                "schema": {
                  "format": "uuid",
                  "type": "string"
                }
              }
            }
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error",
            "headers": {
              "Cache-Control": {
                "description": "Sensitive API responses are private and not stored",
                "schema": {
                  "type": "string"
                }
              },
              "X-Request-ID": {
                "description": "Opaque request correlation identifier",
                "schema": {
                  "format": "uuid",
                  "type": "string"
                }
              }
            }
          }
        },
        "summary": "Logout",
        "tags": [
          "authentication"
        ]
      }
    },
    "/api/v1/auth/session": {
      "get": {
        "operationId": "getSession",
        "parameters": [
          {
            "in": "cookie",
            "name": "session_id",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Session Id"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SessionResponse"
                }
              }
            },
            "description": "Successful Response",
            "headers": {
              "Cache-Control": {
                "description": "Sensitive API responses are private and not stored",
                "schema": {
                  "type": "string"
                }
              },
              "X-Request-ID": {
                "description": "Opaque request correlation identifier",
                "schema": {
                  "format": "uuid",
                  "type": "string"
                }
              }
            }
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error",
            "headers": {
              "Cache-Control": {
                "description": "Sensitive API responses are private and not stored",
                "schema": {
                  "type": "string"
                }
              },
              "X-Request-ID": {
                "description": "Opaque request correlation identifier",
                "schema": {
                  "format": "uuid",
                  "type": "string"
                }
              }
            }
          }
        },
        "summary": "Get Session",
        "tags": [
          "authentication"
        ]
      }
    },
...
  }
}